LPS-71838 Move OSGi commands classes to their own package
