+++ modules/apps/social-office-upgrade/social-office-upgrade/src/main/java/com/liferay/social/office/upgrade/internal/osgi/commands/SocialOfficeUpgradeOSGiCommands.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.social.office.upgrade.internal;
+package com.liferay.social.office.upgrade.internal.osgi.commands;
 
 import com.liferay.portal.kernel.dao.orm.ActionableDynamicQuery;
 import com.liferay.portal.kernel.dao.orm.DynamicQuery;
