+++ modules/post-upgrade-fix/post-upgrade-fix-api/bnd.bnd
@@ -1,6 +1,6 @@
 Bundle-Name: Liferay Post Upgrade Fix API
 Bundle-SymbolicName: com.liferay.post.upgrade.fix.api
 Bundle-Version: 1.0.1
-Export-Package: com.liferay.post.upgrade.fix
+Export-Package: com.liferay.post.upgrade.fix.osgi.commands
 Liferay-Releng-Module-Group-Description:
 Liferay-Releng-Module-Group-Title:
